added exception handling

+++ LMS/src/main/java/com/app/lms/controller/LibraryAdminController.java
@@ -5,8 +5,10 @@ import java.util.Map;
 
 import javax.validation.Valid;
 
+import org.hibernate.TransactionException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.http.converter.HttpMessageNotReadableException;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -16,6 +18,11 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.context.request.RequestContextHolder;
+import org.springframework.web.context.request.ServletRequestAttributes;
+import org.springframework.web.filter.RequestContextFilter;
+import org.springframework.web.servlet.DispatcherServlet;
+
 import com.app.lms.model.LibrarySection;
 import com.app.lms.model.PackageSection;
 import com.app.lms.model.SubscriptionPackage;
@@ -42,6 +49,8 @@ public class LibraryAdminController {
 
 	@GetMapping("/section")
 	public LibrarySection getLibrarySection(@RequestParam("id") String id) {
+		if(id.length()!=3)
+			throw new TransactionException("Invalid id : "+id);
 		return librarySectionService.getLibrarySection(id);
 	}
 
@@ -51,13 +60,18 @@ public class LibraryAdminController {
 	}
 
 	@PutMapping("/section")
-	public void updateLibrarySection(@RequestBody LibrarySection librarySection) {
+	public void updateLibrarySection(@Valid @RequestBody LibrarySection librarySection) {
 		librarySectionService.updateLibrarySection(librarySection);
 	}
 
 	@DeleteMapping("/section/{id}")
-	public void deleteLibrarySection(@PathVariable("id") String id) {
-		librarySectionService.deleteLibrarySection(id);
+	public String deleteLibrarySection(@PathVariable("id") String id) {
+		if(id.length()!=3)
+			throw new TransactionException("Invalid id : "+id);
+		boolean flag = librarySectionService.deleteLibrarySection(id);
+		if(flag)
+			return "Library Section deleted sucessfully";
+		return "Transaction Failed";
 	}
 
 	@PostMapping("/package")
@@ -70,6 +84,8 @@ public class LibraryAdminController {
 		for (JsonNode node : arrayNode) {
 			map.put(node.get("sectionId").asText(), node.get("noOfBooks").asInt());
 		}
+		if(pkg==null || map.isEmpty())
+			throw new TransactionException("Invalid Request");
 		subpkgService.addSubscriptionPackage(pkg, map);
 	}
 
@@ -77,6 +93,8 @@ public class LibraryAdminController {
 	public String getSubscriptionPackage(@RequestParam("id") int id) throws JsonProcessingException {
 		ObjectMapper mapper = new ObjectMapper();
 		SubscriptionPackage pkg = subpkgService.getSubscriptionPackage(id);
+		if(pkg==null)
+			throw new TransactionException("Package not found");
 		ObjectNode objectNode = mapper.valueToTree(pkg);
 		ArrayNode arrayNode = objectNode.putArray("sections");
 		for (PackageSection ps : pkg.getPackageSection()) {
@@ -89,7 +107,10 @@ public class LibraryAdminController {
 	}
 
 	@DeleteMapping("/package/{id}")
-	public void deleteSubscriptionPackage(@PathVariable("id") int id) {
-		subpkgService.deleteSubscriptionPackage(id);
+	public String deleteSubscriptionPackage(@PathVariable("id") int id) {
+		boolean flag = subpkgService.deleteSubscriptionPackage(id);
+		if(flag)
+			return "Subscription Package deleted sucessfully";
+		return "Transaction Failed";
 	}
 }
